refactor: spit steps into separate files
