+++ src/domains/transaction/pages/ResignRegistration/ResignRegistration.tsx
@@ -1,214 +1,20 @@
 import { Contracts } from "@arkecosystem/platform-sdk";
-import { ReadWriteWallet, WalletData } from "@arkecosystem/platform-sdk-profiles";
-import { BigNumber } from "@arkecosystem/platform-sdk-support";
-import { upperFirst } from "@arkecosystem/utils";
-import { Address } from "app/components/Address";
-import { Alert } from "app/components/Alert";
-import { Amount } from "app/components/Amount";
-import { Avatar } from "app/components/Avatar";
+import {  WalletData } from "@arkecosystem/platform-sdk-profiles";
 import { Button } from "app/components/Button";
-import { Circle } from "app/components/Circle";
-import { Form, FormField, FormHelperText,FormLabel } from "app/components/Form";
+import { Form } from "app/components/Form";
 import { Icon } from "app/components/Icon";
-import { InputPassword } from "app/components/Input";
-import { Label } from "app/components/Label";
 import { Page, Section } from "app/components/Layout";
 import { StepIndicator } from "app/components/StepIndicator";
 import { TabPanel, Tabs } from "app/components/Tabs";
-import { TransactionDetail } from "app/components/TransactionDetail";
 import { useEnvironmentContext } from "app/contexts";
 import { useActiveProfile, useActiveWallet } from "app/hooks/env";
-import { InputFee } from "domains/transaction/components/InputFee";
-import { LedgerConfirmation } from "domains/transaction/components/LedgerConfirmation";
-import { TotalAmountBox } from "domains/transaction/components/TotalAmountBox";
-import { TransactionSuccessful } from "domains/transaction/components/TransactionSuccessful";
 import React, { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
 
-type PasswordType = "mnemonic" | "password" | "ledger";
-
-type ResignRegistrationProps = {
-	formDefaultData?: any;
-	onDownload: (transaction: Contracts.SignedTransactionData) => void;
-	passwordType: PasswordType;
-};
-
-type StepProps = {
-	wallet: ReadWriteWallet;
-	delegate: WalletData | any;
-	fee: Contracts.TransactionFee;
-	transaction?: Contracts.SignedTransactionData;
-};
-
-const FirstStep = ({ wallet, delegate, fee }: StepProps) => {
-	const { t } = useTranslation();
-
-	return (
-		<div data-testid="ResignRegistration__first-step">
-			<h1 className="mb-0">{t("TRANSACTION.PAGE_RESIGN_REGISTRATION.FIRST_STEP.DELEGATE.TITLE")}</h1>
-			<div className="text-theme-neutral-dark">
-				{t("TRANSACTION.PAGE_RESIGN_REGISTRATION.FIRST_STEP.DELEGATE.DESCRIPTION")}
-			</div>
-
-			<div className="mt-6">
-				<Alert size="lg">{t("TRANSACTION.PAGE_RESIGN_REGISTRATION.FIRST_STEP.DELEGATE.WARNING")}</Alert>
-			</div>
-
-			<div>
-				<TransactionDetail extra={<Avatar size="lg" address={wallet.address()} />} border={false}>
-					<div className="mb-2 text-sm font-semibold text-theme-neutral">
-						<span className="mr-1">{t("TRANSACTION.SENDER")}</span>
-						<Label color="warning">
-							<span className="text-sm">{t("TRANSACTION.YOUR_ADDRESS")}</span>
-						</Label>
-					</div>
-					<Address address={wallet.address()} />
-				</TransactionDetail>
-
-				<TransactionDetail label={t("TRANSACTION.DELEGATE_NAME")}>{delegate.username()}</TransactionDetail>
-
-				<TransactionDetail className="pt-6 pb-0">
-					<FormField name="name" className="font-normal">
-						<FormLabel>{t("TRANSACTION.TRANSACTION_FEE")}</FormLabel>
-						<InputFee
-							value={fee.static}
-							defaultValue={fee.static}
-							average={fee.avg}
-							min={fee.min}
-							max={fee.max}
-							step={0.01}
-						/>
-					</FormField>
-				</TransactionDetail>
-			</div>
-		</div>
-	);
-};
-
-const SecondStep = ({ wallet, delegate, fee }: StepProps) => {
-	const { t } = useTranslation();
-	const coinName = wallet.manifest().get<string>("name");
-	const network = `${coinName} ${wallet.network().name}`;
-
-	return (
-		<div data-testid="ResignRegistration__second-step">
-			<div>
-				<h1 className="mb-0">{t("TRANSACTION.PAGE_RESIGN_REGISTRATION.SECOND_STEP.TITLE")}</h1>
-				<p className="text-theme-neutral-dark">
-					{t("TRANSACTION.PAGE_RESIGN_REGISTRATION.SECOND_STEP.DESCRIPTION")}
-				</p>
-			</div>
-			<div className="mt-4 grid grid-flow-row gap-2">
-				<TransactionDetail
-					border={false}
-					label={t("TRANSACTION.NETWORK")}
-					extra={
-						<div className="ml-1 text-theme-danger">
-							<Circle className="bg-theme-background border-theme-danger-light" size="lg">
-								{coinName && <Icon name={upperFirst(coinName.toLowerCase())} width={20} height={20} />}
-							</Circle>
-						</div>
-					}
-				>
-					<div className="flex-auto font-semibold truncate text-md text-theme-neutral-800 max-w-24">
-						{network}
-					</div>
-				</TransactionDetail>
-
-				<TransactionDetail extra={<Avatar size="lg" address={wallet.address()} />}>
-					<div className="mb-2 text-sm font-semibold text-theme-neutral">
-						<span className="mr-1">{t("TRANSACTION.SENDER")}</span>
-						<Label color="warning">
-							<span className="text-sm">{t("TRANSACTION.YOUR_ADDRESS")}</span>
-						</Label>
-					</div>
-					<Address address={wallet.address()} />
-				</TransactionDetail>
-
-				<TransactionDetail label={t("TRANSACTION.DELEGATE_NAME")}>{delegate.username()}</TransactionDetail>
-
-				<div className="my-4">
-					<TotalAmountBox amount={BigNumber.ZERO} fee={BigNumber.make(fee.static)} />
-				</div>
-			</div>
-		</div>
-	);
-};
-
-const ThirdStep = ({ form, passwordType }: { form: any; passwordType: PasswordType }) => {
-	const { register } = form;
-
-	const { t } = useTranslation();
-
-	return (
-		<div data-testid="ResignRegistration__third-step">
-			{passwordType !== "ledger" && (
-				<div>
-					<h1 className="mb-0">{t("TRANSACTION.AUTHENTICATION_STEP.TITLE")}</h1>
-					<div className="text-theme-neutral-dark">{t("TRANSACTION.AUTHENTICATION_STEP.DESCRIPTION")}</div>
-
-					<div className="mt-8">
-						<FormField name={passwordType}>
-							<FormLabel>
-								{passwordType === "mnemonic"
-									? t("TRANSACTION.MNEMONIC")
-									: t("TRANSACTION.ENCRYPTION_PASSWORD")}
-							</FormLabel>
-							<InputPassword
-								data-testid={`ResignRegistration__${passwordType}`}
-								name={passwordType}
-								ref={register}
-								onChange={() => form.clearError(passwordType)}
-							/>
-							<FormHelperText />
-						</FormField>
-					</div>
-				</div>
-			)}
-
-			{passwordType === "ledger" && (
-				<div>
-					<h1>{t("TRANSACTION.LEDGER_CONFIRMATION.TITLE")}</h1>
-					<LedgerConfirmation />
-				</div>
-			)}
-		</div>
-	);
-};
-
-export const FourthStep = ({ delegate, fee }: StepProps) => {
-	const { t } = useTranslation();
-
-	return (
-		<TransactionSuccessful>
-			<TransactionDetail
-				label={t("TRANSACTION.TRANSACTION_TYPE")}
-				extra={
-					<Circle className="border-black" size="lg">
-						<Icon name="Delegate" width={20} height={20} />
-					</Circle>
-				}
-			>
-				{t("TRANSACTION.PAGE_RESIGN_REGISTRATION.FOURTH_STEP.TITLE")}
-			</TransactionDetail>
-			<TransactionDetail label={t("TRANSACTION.DELEGATE_NAME")}>{delegate.username()}</TransactionDetail>
-			<TransactionDetail
-				label={t("TRANSACTION.AMOUNT")}
-				extra={
-					<div className="ml-1 text-theme-danger">
-						<Circle className="bg-theme-background border-theme-danger-light" size="lg">
-							<Icon name="Sent" width={22} height={22} />
-						</Circle>
-					</div>
-				}
-			>
-				<Amount ticker="ARK" value={BigNumber.make(fee.static)} />
-			</TransactionDetail>
-		</TransactionSuccessful>
-	);
-};
+import { FirstStep, FourthStep,SecondStep, ThirdStep } from "./";
+import {  ResignRegistrationProps } from "./ResignRegistration.models";
 
 export const ResignRegistration = ({ formDefaultData, onDownload, passwordType }: ResignRegistrationProps) => {
 	const [activeTab, setActiveTab] = useState(1);
