+++ src/domains/transaction/pages/ResignRegistration/index.ts
@@ -1,1 +1,5 @@
 export * from "./ResignRegistration";
+export * from "./Step1";
+export * from "./Step2";
+export * from "./Step3";
+export * from "./Step4";
